song_service/requirements.txt
asgiref==3.5.2
build==0.9.0
click==8.1.3
Django==4.1.2
django-admin==2.0.1
django-cors-headers==3.13.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
djangorestframework==3.14.0
packaging==21.3
pep517==0.13.0
pip-tools==6.9.0
psycopg2-binary==2.9.5
pyparsing==3.0.9
pytz==2022.5
screen==1.0.1
sqlparse==0.4.3
tomli==2.0.1
xlwt==1.3.0
